+++ internal/nodes/crd_translator.go
@@ -15,16 +15,13 @@
 package nodes
 
 import (
-	"encoding/json"
 	"fmt"
-	"reflect"
 	"regexp"
 	"strings"
 
 	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/internal/cache"
 	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/internal/lib"
 	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/internal/objects"
-	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/internal/status"
 	akov1alpha1 "github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/pkg/apis/ako/v1alpha1"
 	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/pkg/utils"
 )
@@ -261,218 +258,4 @@ func BuildPoolHTTPRule(host, path, ingName, namespace, key string, vsNode AviVsE
 	}
 
 	return
-}
-
-// validateHostRuleObj would do validation checks
-// update internal CRD caches, and push relevant ingresses to ingestion
-func validateHostRuleObj(key string, hostrule *akov1alpha1.HostRule) error {
-	var err error
-	fqdn := hostrule.Spec.VirtualHost.Fqdn
-	foundHost, foundHR := objects.SharedCRDLister().GetFQDNToHostruleMapping(fqdn)
-	if foundHost && foundHR != hostrule.Namespace+"/"+hostrule.Name {
-		err = fmt.Errorf("duplicate fqdn %s found in %s", fqdn, foundHR)
-		status.UpdateHostRuleStatus(key, hostrule, status.UpdateCRDStatusOptions{
-			Status: lib.StatusRejected,
-			Error:  err.Error(),
-		})
-		utils.AviLog.Warnf("key: %s, msg: %v", key, err)
-		return err
-	}
-
-	refData := map[string]string{
-		hostrule.Spec.VirtualHost.WAFPolicy:                  "WafPolicy",
-		hostrule.Spec.VirtualHost.ApplicationProfile:         "AppProfile",
-		hostrule.Spec.VirtualHost.TLS.SSLKeyCertificate.Name: "SslKeyCert",
-		hostrule.Spec.VirtualHost.TLS.SSLProfile:             "SslProfile",
-		hostrule.Spec.VirtualHost.AnalyticsProfile:           "AnalyticsProfile",
-		hostrule.Spec.VirtualHost.ErrorPageProfile:           "ErrorPageProfile",
-	}
-
-	for _, policy := range hostrule.Spec.VirtualHost.HTTPPolicy.PolicySets {
-		refData[policy] = "HttpPolicySet"
-	}
-
-	for _, script := range hostrule.Spec.VirtualHost.Datascripts {
-		refData[script] = "VsDatascript"
-	}
-
-	if err := checkRefsOnController(key, refData); err != nil {
-		status.UpdateHostRuleStatus(key, hostrule, status.UpdateCRDStatusOptions{
-			Status: lib.StatusRejected,
-			Error:  err.Error(),
-		})
-		return err
-	}
-
-	status.UpdateHostRuleStatus(key, hostrule, status.UpdateCRDStatusOptions{
-		Status: lib.StatusAccepted,
-		Error:  "",
-	})
-	return nil
-}
-
-// validateHTTPRuleObj would do validation checks
-// update internal CRD caches, and push relevant ingresses to ingestion
-func validateHTTPRuleObj(key string, httprule *akov1alpha1.HTTPRule) error {
-	refData := make(map[string]string)
-	for _, path := range httprule.Spec.Paths {
-		refData[path.TLS.SSLProfile] = "SslProfile"
-		refData[path.ApplicationPersistence] = "ApplicationPersistence"
-
-		for _, hm := range path.HealthMonitors {
-			refData[hm] = "HealthMonitor"
-		}
-	}
-
-	if err := checkRefsOnController(key, refData); err != nil {
-		status.UpdateHTTPRuleStatus(key, httprule, status.UpdateCRDStatusOptions{
-			Status: lib.StatusRejected,
-			Error:  err.Error(),
-		})
-		return err
-	}
-
-	status.UpdateHTTPRuleStatus(key, httprule, status.UpdateCRDStatusOptions{
-		Status: lib.StatusAccepted,
-		Error:  "",
-	})
-	return nil
-}
-
-// validateAviInfraSetting would do validaion checks on the
-// ingested AviInfraSetting objects
-func validateAviInfraSetting(key string, infraSetting *akov1alpha1.AviInfraSetting) error {
-	refData := make(map[string]string)
-	for _, networkName := range infraSetting.Spec.Network.Names {
-		refData[networkName] = "Network"
-	}
-
-	if infraSetting.Spec.SeGroup.Name != "" {
-		refData[infraSetting.Spec.SeGroup.Name] = "ServiceEngineGroup"
-		addSeGroupLabel(key, infraSetting.Spec.SeGroup.Name)
-	}
-
-	if err := checkRefsOnController(key, refData); err != nil {
-		status.UpdateAviInfraSettingStatus(key, infraSetting, status.UpdateCRDStatusOptions{
-			Status: lib.StatusRejected,
-			Error:  err.Error(),
-		})
-		return err
-	}
-
-	status.UpdateAviInfraSettingStatus(key, infraSetting, status.UpdateCRDStatusOptions{
-		Status: lib.StatusAccepted,
-		Error:  "",
-	})
-	return nil
-}
-
-// addSeGroupLabel configures SEGroup with appropriate labels, during AviInfraSetting
-// creation/updates after ingestion
-func addSeGroupLabel(key, segName string) {
-	// assign the last avi client for ref checks
-	clients := cache.SharedAVIClients()
-	aviClientLen := lib.GetshardSize()
-
-	// configure labels on SeGroup if not present already.
-	seGroup, err := cache.GetAviSeGroup(clients.AviClient[aviClientLen], segName)
-	if err != nil {
-		utils.AviLog.Error(err)
-		return
-	}
-	cache.ConfigureSeGroupLabels(clients.AviClient[aviClientLen], seGroup)
-}
-
-var refModelMap = map[string]string{
-	"SslKeyCert":             "sslkeyandcertificate",
-	"WafPolicy":              "wafpolicy",
-	"HttpPolicySet":          "httppolicyset",
-	"SslProfile":             "sslprofile",
-	"AppProfile":             "applicationprofile",
-	"AnalyticsProfile":       "analyticsprofile",
-	"ErrorPageProfile":       "errorpageprofile",
-	"VsDatascript":           "vsdatascriptset",
-	"HealthMonitor":          "healthmonitor",
-	"ApplicationPersistence": "applicationpersistenceprofile",
-	"ServiceEngineGroup":     "serviceenginegroup",
-	"Network":                "network",
-}
-
-func checkRefsOnController(key string, refMap map[string]string) error {
-	for k, value := range refMap {
-		if k == "" {
-			continue
-		}
-
-		if err := checkRefOnController(key, value, k); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// checkRefOnController checks whether a provided ref on the controller
-func checkRefOnController(key, refKey, refValue string) error {
-	uri := fmt.Sprintf("/api/%s?name=%s&fields=name,type,labels,created_by", refModelMap[refKey], refValue)
-	clients := cache.SharedAVIClients()
-
-	// assign the last avi client for ref checks
-	aviClientLen := lib.GetshardSize()
-	if aviClientLen != 0 {
-		aviClientLen = aviClientLen - 1
-	}
-	result, err := lib.AviGetCollectionRaw(clients.AviClient[aviClientLen], uri)
-	if err != nil {
-		utils.AviLog.Warnf("key: %s, msg: Get uri %v returned err %v", key, uri, err)
-		return fmt.Errorf("%s \"%s\" not found on controller", refModelMap[refKey], refValue)
-	}
-
-	if result.Count == 0 {
-		utils.AviLog.Warnf("key: %s, msg: No Objects found for refName: %s/%s", key, refModelMap[refKey], refValue)
-		return fmt.Errorf("%s \"%s\" not found on controller", refModelMap[refKey], refValue)
-	}
-
-	items := make([]json.RawMessage, result.Count)
-	err = json.Unmarshal(result.Results, &items)
-	if err != nil {
-		utils.AviLog.Warnf("key: %s, msg: Failed to unmarshal results, err: %v", key, err)
-		return fmt.Errorf("%s \"%s\" not found on controller", refModelMap[refKey], refValue)
-	}
-
-	item := make(map[string]interface{})
-	err = json.Unmarshal(items[0], &item)
-	if err != nil {
-		utils.AviLog.Warnf("key: %s, msg: Failed to unmarshal item, err: %v", key, err)
-		return fmt.Errorf("%s \"%s\" found on controller is invalid", refModelMap[refKey], refValue)
-	}
-
-	switch refKey {
-	case "AppProfile":
-		if appProfType, ok := item["type"].(string); ok && appProfType != lib.AllowedApplicationProfile {
-			utils.AviLog.Warnf("key: %s, msg: applicationProfile: %s must be of type %s", key, refValue, lib.AllowedApplicationProfile)
-			return fmt.Errorf("%s \"%s\" found on controller is invalid, must be of type: %s",
-				refModelMap[refKey], refValue, lib.AllowedApplicationProfile)
-		}
-	case "ServiceEngineGroup":
-		if seGroupLabels, ok := item["labels"].([]map[string]string); ok {
-			if len(seGroupLabels) == 0 {
-				utils.AviLog.Infof("key: %s, msg: ServiceEngineGroup %s not configured with labels", key, item["name"].(string))
-			} else {
-				if !reflect.DeepEqual(seGroupLabels, lib.GetLabels()) {
-					utils.AviLog.Warnf("key: %s, msg: serviceEngineGroup: %s mismatched labels %s", key, refValue, utils.Stringify(seGroupLabels))
-					return fmt.Errorf("%s \"%s\" found on controller is invalid, mismatched labels: %s",
-						refModelMap[refKey], refValue, utils.Stringify(seGroupLabels))
-				}
-			}
-		}
-	}
-
-	if itemCreatedBy, ok := item["created_by"].(string); ok && itemCreatedBy == lib.GetAKOUser() {
-		utils.AviLog.Warnf("key: %s, msg: Cannot use object referred in CRD created by current AKO instance", key)
-		return fmt.Errorf("%s \"%s\" Invalid operation, object referred is created by current AKO instance",
-			refModelMap[refKey], refValue)
-	}
-
-	utils.AviLog.Infof("key: %s, msg: Ref found for %s/%s", key, refModelMap[refKey], refValue)
-	return nil
 }
