+++ tests/hostnameshardtests/l7_evh_hostnameshard_rest_test.go
@@ -764,7 +764,7 @@ func TestCUDSecretCacheSyncForEvh(t *testing.T) {
 	g.Eventually(func() bool {
 		_, found := mcache.SSLKeyCache.AviCacheGet(sslKey)
 		return found
-	}, 30*time.Second).Should(gomega.Equal(true))
+	}, 50*time.Second).Should(gomega.Equal(true))
 	parentVSCache, _ := mcache.VsCacheMeta.AviCacheGet(parentVSKey)
 	parentVSCacheObj, _ := parentVSCache.(*cache.AviVsCache)
 	g.Expect(parentVSCacheObj.HTTPKeyCollection).To(gomega.HaveLen(0))
